Added some helpful functions.
Added a way for Taskables to access print in printFromObject.
Added functions to loop operations in loopOp and loopOpInf
Added function to get an object's reference from its id. (passReference)
Added a function to get the id of all objects in a session.(fullDiscovery)
Added method append to shift. So shift.append(o) = shift.operations.append(o)
Changed Session and SessionDirectory error logging to exclude StopIteration exceptions.
Changed __str__ of Data to be more readable.
Changed some comments.
Tested stuff that was added.
Doced stuff that was added.
Commented stuff that was added.

Signed-off-by: Jacob Ledbetter <ledbetterte@gmail.com>